requirements.txt
beautifulsoup4
python-dotenv
requests
pendulum
boto3
botocore